Add unit tests for memory fetch and memory creation.
Fetching memories is simply tested by checking that the number of memories retrieved by the query is correct.
Creating new memory is tested by checking that the total number of memories and the number of memories of the corresponding owner increase by 1 after successfully creation. Then the newly created memories is retrieved to check if all values are inserted correctly.